requirements.txt
streamlit
tensorflow
opencv-python-headless
numpy
pandas
Pillow
scikit-learn
joblib